+++ static_frontend/src/components/index.ts
@@ -1,11 +1,9 @@
 import Navbar from './Navbar';
 import Footer from './Footer';
-import Layout from './Layout';
 import AccordionTile from './accordion/AccordionTile';
 
 export {
   Navbar,
   Footer,
-  Layout,
   AccordionTile
 };
